setup.sh fixes

+++ setup.sh
@@ -10,7 +10,9 @@ fi
 PIE_HOME=/home/pi/RetroPie
 MY_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-killall emulationstation
+if ps faux | grep emulationstation | grep -v grep; then
+  killall emulationstation
+fi
 
 # Create savestates/savefiles dirs for each system
 for system in $(find "${PIE_HOME}/roms" -maxdepth 1 -mindepth 1 -type d -exec basename {} \;); do
@@ -24,12 +26,29 @@ chown -R pi:pi "${PIE_HOME}/savestates"
 cd "${MY_DIR}/files"
 for source_file in $(find . -type f); do
   dest_file="${source_file:1}"
+  echo "FILE => ${dest_file}"
+
+  # /boot is special
+  if [[ "${dest_file}" == /boot/* ]]; then
+    cp "${source_file}" "${dest_file}"
+    chown root:root "${dest_file}"
+    continue
+  fi
+
+  # Ensure dest dir exists
+  dir="$(dirname "${dest_file}")"
+  if [ ! -d "${dir}" ]; then
+    mkdir -p "${dir}"
+    chown pi:pi "${dir}"
+  fi
+
   cp "${source_file}" "${dest_file}"
   chown pi:pi "${dest_file}"
 done
 
-if ! grep "[SaveStates]" /etc/samba/smb.conf; then
-  cat <<EOF >> /etc/samba/smb.conf
+if ! grep "[SaveStates]" /etc/samba/smb.conf &> /dev/null; then
+  echo "=> Writing samba config"
+  cat >> /etc/samba/smb.conf <<EOF
 [SaveStates]
 comment = pi
 path = "/home/pi/RetroPie/savestates"
